fix storybook move e2e in content for versioning fix lint
