+++ e2e/content-services/components/viewer-vesion.e2e.ts
@@ -16,10 +16,11 @@
  */
 
 import { browser } from 'protractor';
-import { createApiService, LoginPage, UploadActions, UserModel, UsersActions, ViewerPage } from '@alfresco/adf-testing';
+import { createApiService, FileBrowserUtil, LoginPage, UploadActions, UserModel, UsersActions, ViewerPage } from '@alfresco/adf-testing';
 import { ContentServicesPage } from '../../core/pages/content-services.page';
 import { FileModel } from '../../models/ACS/file.model';
 import { NavigationBarPage } from '../../core/pages/navigation-bar.page';
+import { VersionManagePage } from '../pages/version-manager.page';
 
 describe('Viewer', () => {
 
@@ -32,6 +33,7 @@ describe('Viewer', () => {
     const uploadActions = new UploadActions(apiService);
     const usersActions = new UsersActions(apiService);
 
+    const versionManagePage = new VersionManagePage();
     const acsUser = new UserModel();
     let txtFileUploaded;
 
@@ -72,24 +74,21 @@ describe('Viewer', () => {
         await viewerPage.clickCloseButton();
     });
 
-    it('[C260096] Should the Viewer able to accept a customToolbar', async () => {
-        await viewerPage.clickLeftSidebarButton();
-        await viewerPage.checkLeftSideBarIsDisplayed();
-        await viewerPage.checkToolbarIsDisplayed();
-        await viewerPage.enableCustomToolbar();
-        await viewerPage.checkCustomToolbarIsDisplayed();
-        await viewerPage.disableCustomToolbar();
-    });
-
-    it('[C260097] Should the viewer able to show a custom info-drawer when the sidebarTemplate is set', async () => {
+    it('[C362242] Should the Viewer be able to view a previous version of a file', async () => {
+        await contentServicesPage.versionManagerContent(txtFileInfo.name);
+        await versionManagePage.showNewVersionButton.click();
+        await versionManagePage.uploadNewVersionFile(fileModelVersionTwo.location);
+        await versionManagePage.closeVersionDialog();
+        await contentServicesPage.doubleClickRow(txtFileUploaded.entry.name);
+        await viewerPage.waitTillContentLoaded();
         await viewerPage.clickInfoButton();
-        await viewerPage.checkInfoSideBarIsDisplayed();
-        await viewerPage.clickOnTab('Comments');
-        await viewerPage.checkTabIsActive('Comments');
-        await viewerPage.clickOnTab('Properties');
-        await viewerPage.checkTabIsActive('Properties');
         await viewerPage.clickOnTab('Versions');
-        await viewerPage.checkTabIsActive('Versions');
+        await versionManagePage.viewFileVersion('1.0');
+        await viewerPage.expectUrlToContain('1.0');
     });
 
+    it('[C362265] Should the Viewer be able to download a previous version of a file', async () => {
+        await viewerPage.clickDownloadButton();
+        await FileBrowserUtil.isFileDownloaded(txtFileInfo.name);
+    });
 });
