+++ e2e/content-services/document-list/document-list-selection.e2e.ts
@@ -21,7 +21,7 @@ import { NavigationBarPage } from '../../core/pages/navigation-bar.page';
 import { FolderModel } from '../../models/ACS/folder.model';
 import { browser } from 'protractor';
 import { FileModel } from '../../models/ACS/file.model';
-import { UploadDialogPage } from '../../core/pages/dialog/upload-dialog.page';
+import { UploadDialogPage } from '../pages/upload-dialog.page';
 
 describe('Document List - Selection', () => {
     const loginPage = new LoginPage();
