+++ e2e/content-services/pages/upload-dialog.page.ts
