+++ e2e/content-services/pages/version-manager.page.ts
