+++ e2e/content-services/upload/cancel-upload.e2e.ts
@@ -18,7 +18,7 @@
 import { browser } from 'protractor';
 import { createApiService, LoginPage, UploadActions, UserModel, UsersActions } from '@alfresco/adf-testing';
 import { ContentServicesPage } from '../../core/pages/content-services.page';
-import { UploadDialogPage } from '../../core/pages/dialog/upload-dialog.page';
+import { UploadDialogPage } from '../pages/upload-dialog.page';
 import { UploadTogglesPage } from '../../core/pages/dialog/upload-toggles.page';
 import { FileModel } from '../../models/ACS/file.model';
