+++ e2e/content-services/upload/version-actions.e2e.ts
@@ -28,9 +28,9 @@ import { createApiService,
 import { browser, by, element } from 'protractor';
 import { FileModel } from '../../models/ACS/file.model';
 import { ContentServicesPage } from '../../core/pages/content-services.page';
-import { UploadDialogPage } from '../../core/pages/dialog/upload-dialog.page';
+import { UploadDialogPage } from '../pages/upload-dialog.page';
 import { NavigationBarPage } from '../../core/pages/navigation-bar.page';
-import { VersionManagePage } from '../../core/pages/version-manager.page';
+import { VersionManagePage } from '../pages/version-manager.page';
 
 describe('Version component actions', () => {
