+++ e2e/content-services/upload/version-smoke-tests.e2e.ts
@@ -18,7 +18,7 @@
 import { browser } from 'protractor';
 import { createApiService, LoginPage, UploadActions, UserModel, UsersActions } from '@alfresco/adf-testing';
 import { ContentServicesPage } from '../../core/pages/content-services.page';
-import { VersionManagePage } from '../../core/pages/version-manager.page';
+import { VersionManagePage } from '../pages/version-manager.page';
 import { FileModel } from '../../models/ACS/file.model';
 import { NavigationBarPage } from '../../core/pages/navigation-bar.page';
