+++ lib/core/src/lib/testing/core.story.module.ts
@@ -20,9 +20,11 @@ import { CoreModule } from '../core.module';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { TranslateModule } from '@ngx-translate/core';
 import { provideTranslations } from '../translation/translation.service';
+import { AuthModule } from '../../../src/lib/auth/oidc/auth.module';
 
 @NgModule({
     imports: [
+        AuthModule.forRoot(),
         TranslateModule.forRoot(),
         CoreModule.forRoot(),
         BrowserAnimationsModule
