+++ lib/process-services-cloud/src/lib/form/services/content-cloud-node-selector.service.spec.ts
@@ -61,13 +61,13 @@ describe('ContentCloudNodeSelectorService', () => {
         dialog = TestBed.inject(MatDialog);
 
         showWarningSpy = spyOn(notificationService, 'showWarning');
-        openDialogSpy = spyOn(dialog, 'open').and.returnValue(<any> {
+        openDialogSpy = spyOn(dialog, 'open').and.returnValue( {
             afterOpened: () => of({}),
             afterClosed: () => of({}),
             componentInstance: {
                 error: new Subject<any>()
             }
-        });
+        } as any);
         getNodeSpy = spyOn(service.nodesApi, 'getNode');
     });
